Add post validation for vaccination_procedure_code
