+++ lambda_code/src/models/fhir_practitioner.py
@@ -1,10 +1,6 @@
 """Practitioner FHIR R4B validator"""
 from fhir.resources.R4B.practitioner import Practitioner
-
-from models.utils import (
-    pre_validate_list,
-    pre_validate_string,
-)
+from models.utils.pre_validator_utils import PreValidation
 
 
 class PractitionerValidator:
@@ -21,7 +17,7 @@ class PractitionerValidator:
         """Pre-validate that, if name exists, then it is an array of length 1"""
         try:
             name = values["name"]
-            pre_validate_list(name, "name", defined_length=1)
+            PreValidation.for_list(name, "name", defined_length=1)
         except KeyError:
             pass
 
@@ -36,7 +32,7 @@ class PractitionerValidator:
         """
         try:
             name_given = values["name"][0]["given"]
-            pre_validate_list(
+            PreValidation.for_list(
                 name_given,
                 "name[0].given",
                 defined_length=1,
@@ -56,7 +52,7 @@ class PractitionerValidator:
 
         try:
             name_family = values["name"][0]["family"]
-            pre_validate_string(name_family, "name[0].family")
+            PreValidation.for_string(name_family, "name[0].family")
         except KeyError:
             pass
 
@@ -69,7 +65,9 @@ class PractitionerValidator:
         """
         try:
             identifier = values["identifier"]
-            pre_validate_list(identifier, "identifier", defined_length=1)
+            PreValidation.for_list(
+                identifier, "identifier", defined_length=1
+            )
         except KeyError:
             pass
 
@@ -83,7 +81,9 @@ class PractitionerValidator:
         """
         try:
             identifier_value = values["identifier"][0]["value"]
-            pre_validate_string(identifier_value, "identifier[0].value")
+            PreValidation.for_string(
+                identifier_value, "identifier[0].value"
+            )
         except KeyError:
             pass
 
@@ -97,7 +97,9 @@ class PractitionerValidator:
         """
         try:
             identifier_system = values["identifier"][0]["system"]
-            pre_validate_string(identifier_system, "identifier[0].system")
+            PreValidation.for_string(
+                identifier_system, "identifier[0].system"
+            )
         except KeyError:
             pass
 
@@ -110,7 +112,7 @@ class PractitionerValidator:
         """
         try:
             identifier_type_coding = values["identifier"][0]["type"]["coding"]
-            pre_validate_list(
+            PreValidation.for_list(
                 identifier_type_coding,
                 "identifier[0].type.coding",
                 defined_length=1,
@@ -130,7 +132,7 @@ class PractitionerValidator:
             identifier_type_coding_display = values["identifier"][0]["type"]["coding"][
                 0
             ]["display"]
-            pre_validate_string(
+            PreValidation.for_string(
                 identifier_type_coding_display,
                 "identifier[0].type.coding[0].display",
             )
